Refactor to use MicroHost RunnableProcess
This makes it so that we can now run multiple runnable processes as separate threads in our main app. Makes it possible to do interesting things like running a REST API service along with a scheduler which periodically fires an operation to check the database for aging/cleanup, for example.